#113 modify: 이벤트 카드 컴포넌트 분리
