+++ src/pages/EventList/EventList.jsx
@@ -1,10 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import * as S from './EventListStyle';
+import * as S from './EventList.style';
 import { USER_ID } from '../../constants';
-import { useNavigate } from 'react-router-dom';
 import { axiosInstance } from '../../axios';
-import { useSetRecoilState } from 'recoil';
-import { eventIDState } from '../../recoil/atoms/state';
+import EventCard from './EventCard';
 
 const EventList = () => {
   const [events, setEvents] = useState([]);
@@ -58,47 +56,3 @@ const EventList = () => {
 };
 
 export default EventList;
-
-const EventCard = ({ id, title, poster, startDate, endDate }) => {
-  const setContent = useSetRecoilState(eventIDState);
-  const navigate = useNavigate();
-
-  const handleDetail = () => {
-    setContent(id);
-    navigate(`/event/dashboard`);
-  };
-
-  const attendanceCheck = (event) => {
-    event.stopPropagation();
-    navigate('/attendance/student-id');
-  };
-
-  // 행사 종료시
-  const isEventEnded = () => {
-    const today = new Date();
-    const eventEndDate = new Date(endDate || startDate);
-    return today > eventEndDate;
-  };
-
-  return (
-    <S.CardWrapper onClick={handleDetail}>
-      <S.EventImgWrapper>
-        <S.EventImg src={poster} alt="event_poster" />
-      </S.EventImgWrapper>
-      <S.EventTitle>{title}</S.EventTitle>
-      <S.EventDate>
-        <p>일정</p>
-        <S.CardDay>
-          {endDate ? `${startDate} ~ ${endDate}` : startDate}
-        </S.CardDay>
-      </S.EventDate>
-      <S.CheckButton
-        isEnded={isEventEnded()}
-        onClick={attendanceCheck}
-        disabled={isEventEnded()}
-      >
-        {isEventEnded() ? '행사 종료' : '출석 체크'}
-      </S.CheckButton>
-    </S.CardWrapper>
-  );
-};
